move eval_conditions to API models package
as this is a clearly API model
